+++ libft/includes/libft.h
@@ -28,7 +28,7 @@ typedef struct	s_list
 	struct s_list	*next;
 }				t_list;
 int				ft_atoi(const char *str);
-int				ft_atol(const char *str);
+long long		ft_atol(const char *str);
 void			ft_bzero(void *s, size_t n);
 int				ft_isalpha(int c);
 int				ft_isdigit(int c);
